OM2-355 | Fix items stuck at "created" state
This bug was actually caused by our JWT caching mechanism that was
creating a new PDO connection using "username" twice instead of passing
the password value.

That would cause the connection to be invalidated, therefore the lib
wasn't able to connect, create the table or even insert data into it.

The fix was as simple as passing the "password" as the 3rd PDO param.